Extract storage and retrieval to separate class
Extracted the methods to store and retrieve devices from the SharedPreferences to a separate Util class that can be used by multiple activities. This reduces the amount of excess code and the size of the Activities.
Added some comments and JavaDoc to the DeviceStorageUtils class.
