+++ app/src/main/java/com/leondeklerk/smartcontroller/DeviceEditFragment.java
@@ -4,7 +4,6 @@ import android.app.Activity;
 import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
-import android.content.SharedPreferences.Editor;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -14,21 +13,18 @@ import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 import com.google.android.material.appbar.MaterialToolbar;
 import com.google.android.material.button.MaterialButton;
-import com.google.gson.Gson;
-import com.google.gson.reflect.TypeToken;
 import com.leondeklerk.smartcontroller.devices.SmartDevice;
-import java.lang.reflect.Type;
+import com.leondeklerk.smartcontroller.utils.DeviceStorageUtils;
 import java.util.ArrayList;
 import org.jetbrains.annotations.NotNull;
 
 public class DeviceEditFragment extends Fragment implements View.OnClickListener {
 
-  static final String ARG_FRAG_NUM =
-      "com.leondeklerk.smartcontroller.FRAG_NUM"; // TODO Change to device name
+  static final String ARG_FRAG_NUM = "com.leondeklerk.smartcontroller.FRAG_NUM";
   private Activity context;
   private int devNum;
   private ArrayList<SmartDevice> devices;
-  private SharedPreferences preferences;
+  private DeviceStorageUtils deviceStorageUtils;
 
   @Override
   public View onCreateView(
@@ -43,10 +39,13 @@ public class DeviceEditFragment extends Fragment implements View.OnClickListener
     if (args != null) {
       devNum = args.getInt(ARG_FRAG_NUM);
     } else {
-      devNum = 0;
+      context.finish();
     }
-    preferences = context.getSharedPreferences(getString(R.string.dev_prefs), Context.MODE_PRIVATE);
-    getDevices();
+    SharedPreferences preferences = context
+        .getSharedPreferences(getString(R.string.dev_prefs), Context.MODE_PRIVATE);
+    deviceStorageUtils = new DeviceStorageUtils(preferences);
+
+    devices = deviceStorageUtils.getDevices();
 
     MaterialButton cancelButton = view.findViewById(R.id.edit_cancel);
     cancelButton.setOnClickListener(this);
@@ -59,7 +58,7 @@ public class DeviceEditFragment extends Fragment implements View.OnClickListener
   public void onResume() {
     super.onResume();
     MaterialToolbar toolbar = context.findViewById(R.id.toolbar);
-    toolbar.setTitle("Dev num: " + devNum);
+    toolbar.setTitle(devices.get(devNum).getData().getName());
   }
 
   @Override
@@ -70,7 +69,7 @@ public class DeviceEditFragment extends Fragment implements View.OnClickListener
         return;
       case R.id.edit_delete:
         devices.remove(devNum);
-        storeDevices();
+        deviceStorageUtils.storeDevices(devices);
         Intent resultIntent = new Intent();
         resultIntent.putExtra(MainActivity.EXTRA_DEV_CHANGED, true);
         context.setResult(Activity.RESULT_OK, resultIntent);
@@ -79,25 +78,5 @@ public class DeviceEditFragment extends Fragment implements View.OnClickListener
       default:
         Toast.makeText(context, "Unknown device clicked", Toast.LENGTH_SHORT).show();
     }
-  }
-
-  public void getDevices() {
-    String json = preferences.getString("deviceList", null);
-    if (json != null) {
-      Gson gson = new Gson();
-      Type type = new TypeToken<ArrayList<SmartDevice>>() {
-      }.getType();
-      devices = gson.fromJson(json, type);
-    } else {
-      devices = new ArrayList<>();
-    }
-  }
-
-  public void storeDevices() {
-    Editor prefsEditor = preferences.edit();
-    Gson gson = new Gson();
-    String json = gson.toJson(devices);
-    prefsEditor.putString("deviceList", json);
-    prefsEditor.apply();
   }
 }
